+++ gold.cc
@@ -1,5 +1,5 @@
 #include "headers/items/gold.h"
 
-Gold::Gold(int value) : Item{'G', GOLD}, value{value} {}
+Gold::Gold(float value) : Item{'G', GOLD}, value{value} {}
 
 Gold::~Gold() {}
