+++ headers/player.h
@@ -16,15 +16,15 @@ class Player : public Character {
     private:
         TempPotion* potionEffect;
     protected:
-        virtual int getGoldValue(Gold* g);
-        virtual int getGoldValue(GoldHoard* g); // I'd like to just have GoldHoard inherit from Gold, but then deadly diamond
+        virtual float getGoldValue(Gold* g);
+        virtual float getGoldValue(GoldHoard* g); // I'd like to just have GoldHoard inherit from Gold, but then deadly diamond
         virtual void usePotion(Potion* p);
     public:
         Player(int maxHP, int atk, int def, string race);
         virtual ~Player() = 0;
         
         string race;
-        int gold;
+        float gold;
         bool hasCompass;
         bool hasBarrierSuit;
