+++ headers/playerRaces/dwarf.h
@@ -4,8 +4,8 @@
 #include "../player.h"
 
 class Dwarf : public Player {
-        int getGoldValue(Gold* g) override;
-        int getGoldValue(GoldHoard* g) override;
+        float getGoldValue(Gold* g) override;
+        float getGoldValue(GoldHoard* g) override;
     public:
         Dwarf();
         virtual ~Dwarf();
