+++ orc.cc
@@ -6,10 +6,10 @@ Orc::Orc() : Player{180, 30, 25, "Orc"} {}
 
 Orc::~Orc() {}
 
-int Orc::getGoldValue(Gold* g) {
-    return g->value / 2;
+float Orc::getGoldValue(Gold* g) {
+    return g->value / 2.0;
 }
 
-int Orc::getGoldValue(GoldHoard* g) {
-    return g->value / 2;
+float Orc::getGoldValue(GoldHoard* g) {
+    return g->value / 2.0;
 }
